+++ product/src/main/java/com/epam/esm/security/service/UserDetailsServiceImpl.java
@@ -1,13 +1,18 @@
 package com.epam.esm.security.service;
 
+import com.epam.esm.entity.Role;
 import com.epam.esm.entity.User;
 import com.epam.esm.repository.UserRepository;
-import com.epam.esm.security.model.SecurityUser;
+import com.epam.esm.service.impl.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+
+import java.util.HashSet;
+import java.util.Set;
 
 @Service("userDetailsServiceImpl")
 public class UserDetailsServiceImpl implements UserDetailsService {
@@ -21,8 +26,15 @@ public class UserDetailsServiceImpl implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        User user = userRepository.findByEmail(email).orElseThrow(()->
+        User user = userRepository.findByEmail(email).orElseThrow(() ->
                 new UsernameNotFoundException("User does not exist"));
-        return SecurityUser.fromUser(user);
+        return UserServiceImpl.fromUser(user);
+    }
+
+    public static Set<SimpleGrantedAuthority> getAuthorities(Role role) {
+        SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(role.name());
+        Set<SimpleGrantedAuthority> authorities = new HashSet<>();
+        authorities.add(simpleGrantedAuthority);
+        return authorities;
     }
 }
